+++ internal/provider/data_source_sdwan_cisco_dhcp_server_feature_template.go
@@ -80,7 +80,7 @@ func (d *CiscoDHCPServerFeatureTemplateDataSource) Schema(ctx context.Context, r
 				MarkdownDescription: "The description of the feature template",
 				Computed:            true,
 			},
-			"device_types": schema.ListAttribute{
+			"device_types": schema.SetAttribute{
 				MarkdownDescription: "List of supported device types",
 				ElementType:         types.StringType,
 				Computed:            true,