+++ internal/provider/model_sdwan_cisco_dhcp_server_feature_template.go
@@ -34,7 +34,7 @@ type CiscoDHCPServer struct {
 	TemplateType             types.String                  `tfsdk:"template_type"`
 	Name                     types.String                  `tfsdk:"name"`
 	Description              types.String                  `tfsdk:"description"`
-	DeviceTypes              types.List                    `tfsdk:"device_types"`
+	DeviceTypes              types.Set                     `tfsdk:"device_types"`
 	AddressPool              types.String                  `tfsdk:"address_pool"`
 	AddressPoolVariable      types.String                  `tfsdk:"address_pool_variable"`
 	ExcludeAddresses         types.List                    `tfsdk:"exclude_addresses"`
@@ -336,9 +336,9 @@ func (data CiscoDHCPServer) toBody(ctx context.Context) string {
 
 func (data *CiscoDHCPServer) fromBody(ctx context.Context, res gjson.Result) {
 	if value := res.Get("deviceType"); value.Exists() {
-		data.DeviceTypes = helpers.GetStringList(value.Array())
+		data.DeviceTypes = helpers.GetStringSet(value.Array())
 	} else {
-		data.DeviceTypes = types.ListNull(types.StringType)
+		data.DeviceTypes = types.SetNull(types.StringType)
 	}
 	if value := res.Get("templateDescription"); value.Exists() && value.String() != "" {
 		data.Description = types.StringValue(value.String())