+++ internal/provider/model_sdwan_cisco_ntp_feature_template.go
@@ -35,7 +35,7 @@ type CiscoNTP struct {
 	TemplateType                  types.String                 `tfsdk:"template_type"`
 	Name                          types.String                 `tfsdk:"name"`
 	Description                   types.String                 `tfsdk:"description"`
-	DeviceTypes                   types.List                   `tfsdk:"device_types"`
+	DeviceTypes                   types.Set                    `tfsdk:"device_types"`
 	Master                        types.Bool                   `tfsdk:"master"`
 	MasterVariable                types.String                 `tfsdk:"master_variable"`
 	MasterStratum                 types.Int64                  `tfsdk:"master_stratum"`
@@ -295,9 +295,9 @@ func (data CiscoNTP) toBody(ctx context.Context) string {
 
 func (data *CiscoNTP) fromBody(ctx context.Context, res gjson.Result) {
 	if value := res.Get("deviceType"); value.Exists() {
-		data.DeviceTypes = helpers.GetStringList(value.Array())
+		data.DeviceTypes = helpers.GetStringSet(value.Array())
 	} else {
-		data.DeviceTypes = types.ListNull(types.StringType)
+		data.DeviceTypes = types.SetNull(types.StringType)
 	}
 	if value := res.Get("templateDescription"); value.Exists() && value.String() != "" {
 		data.Description = types.StringValue(value.String())