+++ internal/provider/model_sdwan_cisco_omp_feature_template.go
@@ -35,7 +35,7 @@ type CiscoOMP struct {
 	TemplateType                   types.String                  `tfsdk:"template_type"`
 	Name                           types.String                  `tfsdk:"name"`
 	Description                    types.String                  `tfsdk:"description"`
-	DeviceTypes                    types.List                    `tfsdk:"device_types"`
+	DeviceTypes                    types.Set                     `tfsdk:"device_types"`
 	GracefulRestart                types.Bool                    `tfsdk:"graceful_restart"`
 	GracefulRestartVariable        types.String                  `tfsdk:"graceful_restart_variable"`
 	OverlayAs                      types.Int64                   `tfsdk:"overlay_as"`
@@ -338,9 +338,9 @@ func (data CiscoOMP) toBody(ctx context.Context) string {
 
 func (data *CiscoOMP) fromBody(ctx context.Context, res gjson.Result) {
 	if value := res.Get("deviceType"); value.Exists() {
-		data.DeviceTypes = helpers.GetStringList(value.Array())
+		data.DeviceTypes = helpers.GetStringSet(value.Array())
 	} else {
-		data.DeviceTypes = types.ListNull(types.StringType)
+		data.DeviceTypes = types.SetNull(types.StringType)
 	}
 	if value := res.Get("templateDescription"); value.Exists() && value.String() != "" {
 		data.Description = types.StringValue(value.String())