+++ internal/provider/model_sdwan_cisco_ospf_feature_template.go
@@ -35,7 +35,7 @@ type CiscoOSPF struct {
 	TemplateType                                  types.String                  `tfsdk:"template_type"`
 	Name                                          types.String                  `tfsdk:"name"`
 	Description                                   types.String                  `tfsdk:"description"`
-	DeviceTypes                                   types.List                    `tfsdk:"device_types"`
+	DeviceTypes                                   types.Set                     `tfsdk:"device_types"`
 	RouterId                                      types.String                  `tfsdk:"router_id"`
 	RouterIdVariable                              types.String                  `tfsdk:"router_id_variable"`
 	AutoCostReferenceBandwidth                    types.Int64                   `tfsdk:"auto_cost_reference_bandwidth"`
@@ -763,9 +763,9 @@ func (data CiscoOSPF) toBody(ctx context.Context) string {
 
 func (data *CiscoOSPF) fromBody(ctx context.Context, res gjson.Result) {
 	if value := res.Get("deviceType"); value.Exists() {
-		data.DeviceTypes = helpers.GetStringList(value.Array())
+		data.DeviceTypes = helpers.GetStringSet(value.Array())
 	} else {
-		data.DeviceTypes = types.ListNull(types.StringType)
+		data.DeviceTypes = types.SetNull(types.StringType)
 	}
 	if value := res.Get("templateDescription"); value.Exists() && value.String() != "" {
 		data.Description = types.StringValue(value.String())