+++ internal/provider/model_sdwan_cisco_sig_credentials_feature_template.go
@@ -34,7 +34,7 @@ type CiscoSIGCredentials struct {
 	TemplateType                   types.String `tfsdk:"template_type"`
 	Name                           types.String `tfsdk:"name"`
 	Description                    types.String `tfsdk:"description"`
-	DeviceTypes                    types.List   `tfsdk:"device_types"`
+	DeviceTypes                    types.Set    `tfsdk:"device_types"`
 	ZscalerOrganization            types.String `tfsdk:"zscaler_organization"`
 	ZscalerOrganizationVariable    types.String `tfsdk:"zscaler_organization_variable"`
 	ZscalerPartnerBaseUri          types.String `tfsdk:"zscaler_partner_base_uri"`
@@ -214,9 +214,9 @@ func (data CiscoSIGCredentials) toBody(ctx context.Context) string {
 
 func (data *CiscoSIGCredentials) fromBody(ctx context.Context, res gjson.Result) {
 	if value := res.Get("deviceType"); value.Exists() {
-		data.DeviceTypes = helpers.GetStringList(value.Array())
+		data.DeviceTypes = helpers.GetStringSet(value.Array())
 	} else {
-		data.DeviceTypes = types.ListNull(types.StringType)
+		data.DeviceTypes = types.SetNull(types.StringType)
 	}
 	if value := res.Get("templateDescription"); value.Exists() && value.String() != "" {
 		data.Description = types.StringValue(value.String())