+++ internal/provider/model_sdwan_cisco_system_feature_template.go
@@ -35,7 +35,7 @@ type CiscoSystem struct {
 	TemplateType                      types.String                           `tfsdk:"template_type"`
 	Name                              types.String                           `tfsdk:"name"`
 	Description                       types.String                           `tfsdk:"description"`
-	DeviceTypes                       types.List                             `tfsdk:"device_types"`
+	DeviceTypes                       types.Set                              `tfsdk:"device_types"`
 	Timezone                          types.String                           `tfsdk:"timezone"`
 	TimezoneVariable                  types.String                           `tfsdk:"timezone_variable"`
 	Hostname                          types.String                           `tfsdk:"hostname"`
@@ -978,9 +978,9 @@ func (data CiscoSystem) toBody(ctx context.Context) string {
 
 func (data *CiscoSystem) fromBody(ctx context.Context, res gjson.Result) {
 	if value := res.Get("deviceType"); value.Exists() {
-		data.DeviceTypes = helpers.GetStringList(value.Array())
+		data.DeviceTypes = helpers.GetStringSet(value.Array())
 	} else {
-		data.DeviceTypes = types.ListNull(types.StringType)
+		data.DeviceTypes = types.SetNull(types.StringType)
 	}
 	if value := res.Get("templateDescription"); value.Exists() && value.String() != "" {
 		data.Description = types.StringValue(value.String())