+++ internal/provider/model_sdwan_cisco_vpn_interface_feature_template.go
@@ -35,7 +35,7 @@ type CiscoVPNInterface struct {
 	TemplateType                                       types.String                                     `tfsdk:"template_type"`
 	Name                                               types.String                                     `tfsdk:"name"`
 	Description                                        types.String                                     `tfsdk:"description"`
-	DeviceTypes                                        types.List                                       `tfsdk:"device_types"`
+	DeviceTypes                                        types.Set                                        `tfsdk:"device_types"`
 	InterfaceName                                      types.String                                     `tfsdk:"interface_name"`
 	InterfaceNameVariable                              types.String                                     `tfsdk:"interface_name_variable"`
 	InterfaceDescription                               types.String                                     `tfsdk:"interface_description"`
@@ -2491,9 +2491,9 @@ func (data CiscoVPNInterface) toBody(ctx context.Context) string {
 
 func (data *CiscoVPNInterface) fromBody(ctx context.Context, res gjson.Result) {
 	if value := res.Get("deviceType"); value.Exists() {
-		data.DeviceTypes = helpers.GetStringList(value.Array())
+		data.DeviceTypes = helpers.GetStringSet(value.Array())
 	} else {
-		data.DeviceTypes = types.ListNull(types.StringType)
+		data.DeviceTypes = types.SetNull(types.StringType)
 	}
 	if value := res.Get("templateDescription"); value.Exists() && value.String() != "" {
 		data.Description = types.StringValue(value.String())