+++ internal/provider/model_sdwan_cli_template_feature_template.go
@@ -34,7 +34,7 @@ type CLITemplate struct {
 	TemplateType types.String `tfsdk:"template_type"`
 	Name         types.String `tfsdk:"name"`
 	Description  types.String `tfsdk:"description"`
-	DeviceTypes  types.List   `tfsdk:"device_types"`
+	DeviceTypes  types.Set    `tfsdk:"device_types"`
 	CliConfig    types.String `tfsdk:"cli_config"`
 }
 
@@ -67,9 +67,9 @@ func (data CLITemplate) toBody(ctx context.Context) string {
 
 func (data *CLITemplate) fromBody(ctx context.Context, res gjson.Result) {
 	if value := res.Get("deviceType"); value.Exists() {
-		data.DeviceTypes = helpers.GetStringList(value.Array())
+		data.DeviceTypes = helpers.GetStringSet(value.Array())
 	} else {
-		data.DeviceTypes = types.ListNull(types.StringType)
+		data.DeviceTypes = types.SetNull(types.StringType)
 	}
 	if value := res.Get("templateDescription"); value.Exists() && value.String() != "" {
 		data.Description = types.StringValue(value.String())