+++ internal/provider/resource_sdwan_cedge_aaa_feature_template_test.go
@@ -82,11 +82,13 @@ func TestAccSdwanCEdgeAAAFeatureTemplate(t *testing.T) {
 					resource.TestCheckResourceAttr("sdwan_cedge_aaa_feature_template.test", "accounting_rules.0.method", "exec"),
 					resource.TestCheckResourceAttr("sdwan_cedge_aaa_feature_template.test", "accounting_rules.0.privilege_level", "15"),
 					resource.TestCheckResourceAttr("sdwan_cedge_aaa_feature_template.test", "accounting_rules.0.start_stop", "true"),
+					resource.TestCheckResourceAttr("sdwan_cedge_aaa_feature_template.test", "accounting_rules.0.groups", "GROUP1"),
 					resource.TestCheckResourceAttr("sdwan_cedge_aaa_feature_template.test", "authorization_console", "true"),
 					resource.TestCheckResourceAttr("sdwan_cedge_aaa_feature_template.test", "authorization_config_commands", "true"),
 					resource.TestCheckResourceAttr("sdwan_cedge_aaa_feature_template.test", "authorization_rules.0.name", "RULE1"),
 					resource.TestCheckResourceAttr("sdwan_cedge_aaa_feature_template.test", "authorization_rules.0.method", "commands"),
 					resource.TestCheckResourceAttr("sdwan_cedge_aaa_feature_template.test", "authorization_rules.0.privilege_level", "15"),
+					resource.TestCheckResourceAttr("sdwan_cedge_aaa_feature_template.test", "authorization_rules.0.groups", "GROUP1"),
 					resource.TestCheckResourceAttr("sdwan_cedge_aaa_feature_template.test", "authorization_rules.0.authenticated", "true"),
 				),
 			},
@@ -171,7 +173,7 @@ func testAccSdwanCEdgeAAAFeatureTemplateConfig_all() string {
 			method = "exec"
 			privilege_level = "15"
 			start_stop = true
-			group = ["GROUP1"]
+			groups = "GROUP1"
 		}]
 		authorization_console = true
 		authorization_config_commands = true
@@ -179,7 +181,7 @@ func testAccSdwanCEdgeAAAFeatureTemplateConfig_all() string {
 			name = "RULE1"
 			method = "commands"
 			privilege_level = "15"
-			group = ["GROUP1"]
+			groups = "GROUP1"
 			authenticated = true
 		}]
 	}